# api/openapi.yaml
...
openapi: 3.0.3
info:
  title: Ultima Management API
  version: 0.6.0
  description: |
    Management surface of the Ultima server (a Redis superset clone).
    JWT bearer auth (§9.3) applies to every /api/v1 route except
    /auth/login and /auth/refresh when auth.enabled is set. /metrics is
    guarded by an IP allowlist instead of JWT.
paths:
...
  /api/v1/auth/login:
    post:
      operationId: login
      summary: Password (+ optional TOTP) login (public)
      security: []
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Token pair
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenPair'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/auth/refresh:
    post:
      operationId: refresh
      summary: Rotate a refresh token (public)
      security: []
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
      responses:
        '200':
          description: New token pair (old refresh token is consumed)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenPair'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/v1/clients/{id}/kill:
    post:
      operationId: killClient
      summary: Close a client connection
      tags: [server]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: uint64
      responses:
        '200':
          description: Connection closed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: The client cannot be killed from this front-end
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/v1/slowlog:
    get:
      operationId: getSlowlog
      summary: Recent commands slower than slowlog-log-slower-than
      tags: [server]
      parameters:
        - name: count
          in: query
          schema:
            type: integer
            minimum: 1
          description: Maximum entries to return (default slowlog-max-len)
      responses:
        '200':
          description: Slowlog entries, newest first
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SlowlogList'
    delete:
      operationId: resetSlowlog
      summary: Clear the slowlog (SLOWLOG RESET)
      tags: [server]
      responses:
        '200':
          description: Cleared
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
...
  /api/v1/config:
    get:
      operationId: getConfig
      summary: Runtime configuration (CONFIG GET * equivalent)
      tags: [server]
      responses:
        '200':
          description: Parameter name → value map
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConfigMap'
    put:
      operationId: putConfig
      summary: Update runtime configuration (CONFIG SET equivalent)
      tags: [server]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConfigUpdateRequest'
      responses:
        '200':
          description: All entries applied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
        '400':
          description: Bad body, unknown parameter, or rejected value
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/v1/flushdb:
    post:
      operationId: flushdb
      summary: Flush one logical DB (or all)
      tags: [data]
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FlushRequest'
      responses:
        '200':
          description: Flushed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
        '400':
          $ref: '#/components/responses/BadRequest'
...
  /api/v1/bgrewriteaof:
    post:
      operationId: bgrewriteaof
      summary: Rewrite the AOF (BGREWRITEAOF)
      tags: [persistence]
      responses:
        '200':
          description: Rewrite started (or scheduled)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
        '409':
          description: Another background job is active
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    BadRequest:
      description: Malformed request body or parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Missing/invalid token, bad credentials, or TOTP required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Resource does not exist
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
    Status:
      type: object
      required: [status]
      properties:
        status:
          type: string
          example: ok
    Error:
      type: object
      required: [status, error]
      properties:
        status:
          type: string
          example: error
        error:
          type: string
...
    LoginRequest:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
          minLength: 1
          x-oapi-codegen-extra-tags:
            validate: required
        password:
          type: string
          minLength: 1
          x-oapi-codegen-extra-tags:
            validate: required
        totp:
          type: string
          description: Current 6-digit code; required when the account has TOTP on
    RefreshRequest:
      type: object
      required: [refresh_token]
      properties:
        refresh_token:
          type: string
          minLength: 1
          x-oapi-codegen-extra-tags:
            validate: required
...
    TokenPair:
      type: object
      required: [access_token, refresh_token, expires_in]
      properties:
        access_token:
          type: string
        refresh_token:
          type: string
        expires_in:
          type: integer
          format: int64
          description: Access-token lifetime in seconds
...
    SlowlogEntry:
      type: object
      required: [id, timestamp, duration_us, args, client_addr, client_name]
      properties:
        id:
          type: integer
          format: int64
        timestamp:
          type: integer
          format: int64
          description: Unix seconds
        duration_us:
          type: integer
          format: int64
        args:
          type: array
          items:
            type: string
        client_addr:
          type: string
        client_name:
          type: string
    SlowlogList:
      type: object
      required: [entries]
      properties:
        entries:
          type: array
          items:
            $ref: '#/components/schemas/SlowlogEntry'
...
    ConfigMap:
      type: object
      required: [entries]
      properties:
        entries:
          type: object
          additionalProperties:
            type: string
    ConfigUpdateRequest:
      type: object
      required: [entries]
      properties:
        entries:
          type: object
          additionalProperties:
            type: string
          minProperties: 1
    FlushRequest:
      type: object
      properties:
        db:
          type: integer
          minimum: 0
          default: 0
        all:
          type: boolean
          default: false
          description: Flush every logical DB (FLUSHALL)
...
security:
  - bearerAuth: []